# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateStaffTransaction:
    properties:
      amount:
        type: integer
      sales_id:
        type: string
      source_type:
        type: string
      staff_id:
        type: string
      text:
        type: string
      type:
        type: string
    type: object
...
  models.UpdateStaffTransaction:
    properties:
      amount:
        type: integer
      id:
        type: string
      sales_id:
        type: string
      source_type:
        type: string
      staff_id:
        type: string
      text:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /staffTransaction:
    get:
      consumes:
      - application/json
      description: Get List StaffTransaction
      operationId: get_list_staffTransaction
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List StaffTransaction
      tags:
      - StaffTransaction
    post:
      consumes:
      - application/json
      description: Create StaffTransaction
      operationId: create_staffTransaction
      parameters:
      - description: CreateStaffTransactionRequest
        in: body
        name: StaffTransaction
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTransaction'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create StaffTransaction
      tags:
      - StaffTransaction
  /staffTransaction/{id}:
    delete:
      consumes:
      - application/json
      description: Delete StaffTransaction
      operationId: delete_staffTransaction
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete StaffTransaction
      tags:
      - StaffTransaction
    get:
      consumes:
      - application/json
      description: Get By ID StaffTransaction
      operationId: get_by_id_staffTransaction
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID StaffTransaction
      tags:
      - StaffTransaction
    put:
      consumes:
      - application/json
      description: Update StaffTransaction
      operationId: update_staffTransaction
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateStaffTransactionRequest
        in: body
        name: StaffTransaction
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStaffTransaction'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update StaffTransaction
      tags:
      - StaffTransaction
swagger: "2.0"